fix deprecated numpy and scipy calls

+++ shesha/util/kl_util.py
@@ -186,7 +186,7 @@ def radii(nr: int, npp: int, cobs: float) -> np.ndarray:
             r
     """
 
-    r2 = cobs**2 + (np.arange(nr, dtype=np.float) + 0.) / nr * (1.0 - cobs**2)
+    r2 = cobs**2 + (np.arange(nr, dtype=np.float32) + 0.) / nr * (1.0 - cobs**2)
     rs = np.sqrt(r2)
     r = np.transpose(np.tile(rs, (npp, 1)))
 
@@ -215,7 +215,7 @@ def polang(r: np.ndarray) -> np.ndarray:
     s = r.shape
     nr = s[0]
     np1 = s[1]
-    phi1 = np.arange(np1, dtype=np.float) / float(np1) * 2. * np.pi
+    phi1 = np.arange(np1, dtype=np.float32) / float(np1) * 2. * np.pi
     p1, p2 = np.meshgrid(np.ones(nr), phi1)
     p = np.transpose(p2)
 
@@ -348,7 +348,7 @@ def pcgeom(nr, npp, cobs, ncp, ncmar):
     """
     nused = ncp - 2 * ncmar
     ff = 0.5 * nused
-    hw = np.float(ncp - 1) / 2.
+    hw = np.float32(ncp - 1) / 2.
 
     r = radii(nr, npp, cobs)
     p = polang(r)
@@ -358,7 +358,7 @@ def pcgeom(nr, npp, cobs, ncp, ncmar):
     px = ff * px0 + hw
     py = ff * py0 + hw
     ax = np.reshape(
-            np.arange(int(ncp)**2, dtype=np.float) + 1, (int(ncp), int(ncp)), order='F')
+            np.arange(int(ncp)**2, dtype=np.float32) + 1, (int(ncp), int(ncp)), order='F')
     ax = np.float32(ax - 1) % ncp - 0.5 * (ncp - 1)
     ax = ax / (0.5 * nused)
     ay = np.transpose(ax)
